tools/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
enum34==1.1.6
idna==2.6
ipaddress==1.0.19
Js2Py==0.59
paramiko==2.4.0
pyasn1==0.4.2
pycparser==2.18
pyjsparser==2.5.2
PyNaCl==1.2.1
pypac==0.6.0
PySocks==1.6.8
pytz==2018.3
requests==2.18.4
six==1.11.0
sshtunnel==0.1.3
tld==0.7.9
tzlocal==1.5.1
urllib3==1.22
win-inet-pton==1.0.1
